changed zipped logic

+++ src/Constants.java
@@ -1,8 +1,8 @@
 public class Constants {
-    public static int COMPANIES = 3000;
-    public static int PERCISION = 3;
-    public static int WRITE_LINE_LIMIT = 2000;
-    public static int WRITE_FILE_LIMIT = 30;
+    public static short COMPANIES = 3000;
+    public static short PERCISION = 3;
+    public static short WRITE_LINE_LIMIT = 2000;
+    public static short WRITE_FILE_LIMIT = 30;
     public static int ZIP_FILE_DELAY = 1000*60*30;
     public static int START = 0;
     public static int END = 4;
